+++ app/javascript/lca/utils/calculated/ratings/hardness.js
@@ -10,19 +10,21 @@ export function hardness(character: Character): Pool {
   const type = (character.exalt_type || '').toLowerCase()
   const caste = (character.caste || '').toLowerCase()
 
+  armor = character.bonus_hardness
+
   if (character.armor_is_artifact) {
     switch (character.armor_weight) {
       case 'light':
-        armor = 4 + character.bonus_hardness
+        armor += 4
         break
       case 'medium':
-        armor = 7 + character.bonus_hardness
+        armor += 7
         break
       case 'heavy':
-        armor = 10 + character.bonus_hardness
+        armor += 10
         break
       default:
-        armor = character.bonus_hardness
+        break
     }
   }